fix: Heading ID links now work in every case Fixes #3371
